+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - PASSBOLT-2972: As LU I should be able to delete multiple passwords in bulk
 - PASSBOLT-2951: Merge the remember me on CE
 - PASSBOLT-2329: As an administrator deleting a group which is sole owner of one or several passwords, I should be requested to select a new owner for these passwords
+- PASSBOLT-2983: As LU I should be able to share multiple passwords in bulk
 
 ### Improved
 - GITHUB#275: Adding SSL configuration environment variables for cake mysql driver